tests/api/requirements.txt
requests>=2.31.0
pytest>=7.4.0
pytest-xdist>=3.3.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
PyYAML>=6.0